Move game functions to their own file
