+++ Makefile
@@ -1,9 +1,9 @@
 CC=gcc
-CFLAGS=-pipe -O2 -std=c11 -g -Wall -Wextra 
+CFLAGS=-pipe -O2 -std=c11 -g -Wall -Wextra
 LDFLAGS=-I. $(shell pkg-config --cflags --libs sdl2 SDL2_image)
 
-DEPS = src/images.h src/fonts.h
-OBJ = src/netpong.o src/images.o src/fonts.o
+DEPS = src/images.h src/fonts.h src/game.h
+OBJ = src/netpong.o src/images.o src/fonts.o src/game.o
 
 EXECUTABLES=netpong
