+++ src/config.h
@@ -5,6 +5,8 @@
 #define SCREEN_WIDTH 800
 #define SCREEN_HEIGHT 600
 
+#define MAX_TEXT_BUFF_SIZE 100
+
 #define BASE_BALL_SPEED 6
 #define BASE_PADDLE_SPEED 4
 #define CENTER_X (SCREEN_WIDTH / 2)
