+++ src/game.c
@@ -1,0 +1,78 @@
+#include "game.h"
+
+struct game_t Game = {
+    SDL_FALSE,
+    {
+        SCREEN_WIDTH,
+        SCREEN_HEIGHT,
+        SCREEN_NAME,
+        NULL,
+        NULL
+    },
+    game_init,
+    game_quit,
+    0, 0
+};
+
+void init_images()
+{
+    int flags = IMG_INIT_PNG;
+    int rc = IMG_Init(flags);
+
+    if ((rc&flags) != flags) {
+        fprintf(stderr, "SDL error -> %s\n", IMG_GetError());
+        exit(1);
+    }
+}
+
+void
+game_init()
+{
+    int rc = SDL_Init(SDL_INIT_EVERYTHING);
+
+    if (rc != 0) {
+        fprintf(stderr, "SDL error -> %s\n", SDL_GetError());
+        exit(1);
+    }
+
+    init_images();
+
+    Game.screen.window = SDL_CreateWindow("NetPong",
+            SDL_WINDOWPOS_UNDEFINED,
+            SDL_WINDOWPOS_UNDEFINED,
+            Game.screen.width,
+            Game.screen.height,
+            SDL_WINDOW_SHOWN);
+
+    if (Game.screen.window == NULL) {
+        fprintf(stderr, "Window couldn't be created");
+        game_quit();
+        exit(1);
+    }
+
+    Game.screen.renderer = SDL_CreateRenderer(
+        Game.screen.window,
+        -1,
+        SDL_RENDERER_ACCELERATED|SDL_RENDERER_PRESENTVSYNC
+    );
+
+    Game.running = SDL_TRUE;
+
+}
+
+void game_quit()
+{
+    if (Game.screen.renderer) {
+        SDL_DestroyRenderer(Game.screen.renderer);
+    }
+
+    if (Game.screen.window) {
+        SDL_DestroyWindow(Game.screen.window);
+    }
+
+    IMG_Quit();
+    SDL_Quit();
+    Game.running = SDL_FALSE;
+}
+
+
