+++ src/netpong.c
@@ -11,72 +11,8 @@
 #include "images.h"
 #include "fonts.h"
 
-#define MAX_TEXT_BUFF_SIZE 100
-
-struct direction_vector {
-    int x;
-    int y;
-};
-
-void init_images()
-{
-    int flags = IMG_INIT_PNG;
-    int rc = IMG_Init(flags);
-
-    if ((rc&flags) != flags) {
-        fprintf(stderr, "SDL error -> %s\n", IMG_GetError());
-        exit(1);
-    }
-}
-
-void game_init()
-{
-    int rc = SDL_Init(SDL_INIT_EVERYTHING);
-
-    if (rc != 0) {
-        fprintf(stderr, "SDL error -> %s\n", SDL_GetError());
-        exit(1);
-    }
-
-    init_images();
-
-
-    Game.screen.window = SDL_CreateWindow("NetPong",
-            SDL_WINDOWPOS_UNDEFINED,
-            SDL_WINDOWPOS_UNDEFINED,
-            Game.screen.width,
-            Game.screen.height,
-            SDL_WINDOW_SHOWN);
-
-    if (Game.screen.window == NULL) {
-        fprintf(stderr, "Window couldn't be created");
-        game_quit();
-        exit(1);
-    }
-
-    Game.screen.renderer = SDL_CreateRenderer(
-        Game.screen.window,
-        -1,
-        SDL_RENDERER_ACCELERATED|SDL_RENDERER_PRESENTVSYNC
-    );
-
-    Game.running = SDL_TRUE;
-}
-
-void game_quit()
-{
-    if (Game.screen.renderer) {
-        SDL_DestroyRenderer(Game.screen.renderer);
-    }
-
-    if (Game.screen.window) {
-        SDL_DestroyWindow(Game.screen.window);
-    }
-
-    IMG_Quit();
-    SDL_Quit();
-    Game.running = SDL_FALSE;
-}
+const Uint32 FRAMES_PER_SECOND = 60;
+extern struct game_t Game;
 
 void check_events(const Uint8 *keyboardState, int *moving)
 {
@@ -222,7 +158,7 @@ int main()
     Uint32 last_tick = SDL_GetTicks();
     char buffer[100];
 
-    Game.init();
+    game_init();
 
     load_font(Game.screen.renderer, "images/good_neighbors.png", "data/font.txt");
